+++ src/Command/ImportCommand.php
@@ -26,15 +26,14 @@ use DatabaseBackup\Utility\BackupImport;
 use Exception;
 
 /**
- * Imports a database backup
+ * Command to import a database backup.
+ *
  * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-commands#import
  */
 class ImportCommand extends Command
 {
     /**
-     * Hook method for defining this command's option parser
-     * @param \Cake\Console\ConsoleOptionParser $parser The parser to be defined
-     * @return \Cake\Console\ConsoleOptionParser
+     * @inheritDoc
      */
     protected function buildOptionParser(ConsoleOptionParser $parser): ConsoleOptionParser
     {
@@ -45,13 +44,18 @@ class ImportCommand extends Command
                 'required' => true,
             ])
             ->addOption('timeout', [
-                'help' => __d('database_backup', 'Timeout for shell commands. Default value: {0} seconds', Configure::readOrFail('DatabaseBackup.processTimeout')),
+                'help' => __d(
+                    'database_backup',
+                    'Timeout for shell commands. Default value: {0} seconds',
+                    Configure::readOrFail('DatabaseBackup.processTimeout')
+                ),
                 'short' => 't',
             ]);
     }
 
     /**
-     * Internal method to get a `BackupImport` instance
+     * Internal method to get a `BackupImport` instance.
+     *
      * @return \DatabaseBackup\Utility\BackupImport
      */
     protected function getBackupImport(): BackupImport
@@ -60,11 +64,11 @@ class ImportCommand extends Command
     }
 
     /**
-     * Imports a database backup
+     * Imports a database backup.
+     *
      * @param \Cake\Console\Arguments $args The command arguments
      * @param \Cake\Console\ConsoleIo $io The console io
      * @return void
-     * @throws \ReflectionException
      */
     public function execute(Arguments $args, ConsoleIo $io): void
     {
@@ -82,7 +86,9 @@ class ImportCommand extends Command
 
             $file = $BackupImport->import();
             if (!$file) {
-                throw new StopException(__d('database_backup', 'The `{0}` event stopped the operation', 'Backup.beforeImport'));
+                throw new StopException(
+                    __d('database_backup', 'The `{0}` event stopped the operation', 'Backup.beforeImport')
+                );
             }
             $io->success(__d('database_backup', 'Backup `{0}` has been imported', rtr($file)));
         } catch (Exception $e) {