+++ src/Driver/Mysql.php
@@ -19,7 +19,7 @@ namespace DatabaseBackup\Driver;
 use Symfony\Component\Filesystem\Filesystem;
 
 /**
- * Mysql driver to export/import database backups
+ * Mysql driver to export/import database backups.
  */
 class Mysql extends AbstractDriver
 {
@@ -33,16 +33,22 @@ class Mysql extends AbstractDriver
      * Internal method to get the auth file path.
      *
      * This method returns only the path that will be used and does not verify that the file already exists.
+     *
      * @return string
      * @since 2.11.0
      */
     protected function getAuthFilePath(): string
     {
-        return $this->auth ??= TMP . uniqid('auth');
+        if (empty($this->auth)) {
+            $this->auth = TMP . uniqid('auth');
+        }
+
+        return $this->auth;
     }
 
     /**
-     * Internal method to write an auth file
+     * Internal method to write an auth file.
+     *
      * @param string $content Content
      * @return bool
      * @since 2.3.0
@@ -50,8 +56,16 @@ class Mysql extends AbstractDriver
     protected function writeAuthFile(string $content): bool
     {
         $content = str_replace(
-            ['{{USER}}', '{{PASSWORD}}', '{{HOST}}'],
-            [(string)$this->getConfig('username'), (string)$this->getConfig('password'), (string)$this->getConfig('host')],
+            [
+                '{{USER}}',
+                '{{PASSWORD}}',
+                '{{HOST}}',
+            ],
+            [
+                (string)$this->getConfig('username'),
+                (string)$this->getConfig('password'),
+                (string)$this->getConfig('host'),
+            ],
             $content
         );
 
@@ -62,7 +76,8 @@ class Mysql extends AbstractDriver
     }
 
     /**
-     * Called after export
+     * Called after export.
+     *
      * @return void
      * @since 2.1.0
      */
@@ -72,7 +87,8 @@ class Mysql extends AbstractDriver
     }
 
     /**
-     * Called after import
+     * Called after import.
+     *
      * @return void
      * @since 2.1.0
      */
@@ -89,6 +105,7 @@ class Mysql extends AbstractDriver
      * For security reasons, it's recommended to specify the password in a configuration file and not in the command (a
      * user can execute a `ps aux | grep mysqldump` and see the password).
      * So it creates a temporary file to store the configuration options.
+     *
      * @return bool
      * @since 2.1.0
      */
@@ -108,6 +125,7 @@ class Mysql extends AbstractDriver
      * For security reasons, it's recommended to specify the password in a configuration file and not in the command (a
      * user can execute a `ps aux | grep mysqldump` and see the password).
      * So it creates a temporary file to store the configuration options.
+     *
      * @return bool
      * @since 2.1.0
      */
@@ -120,7 +138,8 @@ class Mysql extends AbstractDriver
     }
 
     /**
-     * Deletes the temporary file with the database authentication data
+     * Deletes the temporary file with the database authentication data.
+     *
      * @return void
      * @since 2.1.0
      */