+++ src/Utility/BackupExport.php
@@ -21,7 +21,8 @@ use LogicException;
 use Symfony\Component\Filesystem\Filesystem;
 
 /**
- * Utility to export databases
+ * Utility to export databases.
+ *
  * @property ?string $compression
  * @property ?string $emailRecipient
  * @property string $extension
@@ -30,31 +31,26 @@ use Symfony\Component\Filesystem\Filesystem;
 class BackupExport extends AbstractBackupUtility
 {
     /**
-     * Compression type
      * @var string|null
      */
     protected ?string $compression = null;
 
     /**
-     * Default extension
      * @var string
      */
     private string $defaultExtension = 'sql';
 
     /**
-     * Recipient of the email, if you want to send the backup via mail
      * @var string|null
      */
     protected ?string $emailRecipient = null;
 
     /**
-     * Filename extension
      * @var string
      */
     protected string $extension;
 
     /**
-     * Rotate limit. This is the number of backups you want to keep. So, it will delete all backups that are older
      * @var int
      */
     protected int $rotate = 0;
@@ -66,13 +62,13 @@ class BackupExport extends AbstractBackupUtility
      *  - `bzip2`;
      *  - `gzip`;
      *  - `null` for no compression.
+     *
      * @param string|null $compression Compression type name
-     * @return $this
-     * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupExport-utility#compression
+     * @return self
      * @throws \LogicException
-     * @noinspection PhpMissingReturnTypeInspection
+     * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupExport-utility#compression
      */
-    public function compression(?string $compression)
+    public function compression(?string $compression): self
     {
         $this->extension = $this->defaultExtension;
 
@@ -91,13 +87,13 @@ class BackupExport extends AbstractBackupUtility
      * Sets the filename.
      *
      * The compression type will be automatically set by the filename.
+     *
      * @param string $filename Filename. It can be an absolute path and may contain patterns
-     * @return $this
+     * @return self
      * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupExport-utility#filename
      * @throws \LogicException
-     * @noinspection PhpMissingReturnTypeInspection
      */
-    public function filename(string $filename)
+    public function filename(string $filename): self
     {
         $config = $this->getConnection()->config();
 
@@ -111,13 +107,19 @@ class BackupExport extends AbstractBackupUtility
 
         $filename = $this->getAbsolutePath($filename);
         if (!is_writable(dirname($filename))) {
-            throw new LogicException(__d('database_backup', 'File or directory `{0}` is not writable', dirname($filename)));
+            throw new LogicException(
+                __d('database_backup', 'File or directory `{0}` is not writable', dirname($filename))
+            );
         }
         if (file_exists($filename)) {
-            throw new LogicException(__d('database_backup', 'File `{0}` already exists', $filename));
+            throw new LogicException(
+                __d('database_backup', 'File `{0}` already exists', $filename)
+            );
         }
         if (!$this->getExtension($filename)) {
-            throw new LogicException(__d('database_backup', 'Invalid `{0}` file extension', pathinfo($filename, PATHINFO_EXTENSION)));
+            throw new LogicException(
+                __d('database_backup', 'Invalid `{0}` file extension', pathinfo($filename, PATHINFO_EXTENSION))
+            );
         }
 
         //Sets the compression
@@ -129,13 +131,13 @@ class BackupExport extends AbstractBackupUtility
     }
 
     /**
-     * Sets the number of backups you want to keep. So, it will delete all backups that are older
+     * Sets the number of backups you want to keep. So, it will delete all backups that are older.
+     *
      * @param int $rotate Number of backups you want to keep
-     * @return $this
+     * @return self
      * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupExport-utility#rotate
-     * @noinspection PhpMissingReturnTypeInspection
      */
-    public function rotate(int $rotate)
+    public function rotate(int $rotate): self
     {
         $this->rotate = $rotate;
 
@@ -143,14 +145,20 @@ class BackupExport extends AbstractBackupUtility
     }
 
     /**
-     * Sets the recipient's email address to send the backup file via mail
+     * Sets the recipient's email address to send the backup file via mail.
+     *
      * @param string|null $recipient Recipient's email address or `null` to disable
-     * @return $this
+     * @return self
      * @since 1.1.0
-     * @noinspection PhpMissingReturnTypeInspection
+     * @deprecated 2.13.4 the `BackupExport::send()` method is deprecated. Will be removed in a future release
      */
-    public function send(?string $recipient = null)
+    public function send(?string $recipient = null): self
     {
+        deprecationWarning(
+            '2.13.4',
+            'The `BackupExport::send()` method is deprecated. Will be removed in a future release'
+        );
+
         $this->emailRecipient = $recipient;
 
         return $this;
@@ -162,9 +170,9 @@ class BackupExport extends AbstractBackupUtility
      * When exporting, this method will trigger these events (implemented by the driver instance):
      *  - `Backup.beforeExport`: will be triggered before export;
      *  - `Backup.afterExport`: will be triggered after export.
+     *
      * @return string|false Filename path on success or `false` if the `Backup.beforeExport` event is stopped
      * @throws \LogicException
-     * @throws \ReflectionException
      * @see \DatabaseBackup\Driver\AbstractDriver::afterExport()
      * @see \DatabaseBackup\Driver\AbstractDriver::beforeExport()
      * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupExport-utility#export
@@ -189,7 +197,9 @@ class BackupExport extends AbstractBackupUtility
         //Exports
         $Process = $this->getProcess($this->getDriver()->getExportExecutable($filename));
         if (!$Process->isSuccessful()) {
-            throw new LogicException(__d('database_backup', 'Export failed with error message: `{0}`', rtrim($Process->getErrorOutput())));
+            throw new LogicException(
+                __d('database_backup', 'Export failed with error message: `{0}`', rtrim($Process->getErrorOutput()))
+            );
         }
         (new Filesystem())->chmod($filename, Configure::read('DatabaseBackup.chmod'));