+++ tests/TestCase/Command/DeleteAllCommandTest.php
@@ -17,6 +17,7 @@ namespace DatabaseBackup\Test\TestCase\Command;
 
 use Cake\Console\TestSuite\ConsoleIntegrationTestTrait;
 use DatabaseBackup\TestSuite\TestCase;
+use PHPUnit\Framework\Attributes\WithoutErrorHandler;
 
 /**
  * DeleteAllCommandTest class
@@ -29,7 +30,6 @@ class DeleteAllCommandTest extends TestCase
 
     /**
      * @test
-     * @throws \ReflectionException
      * @uses \DatabaseBackup\Command\DeleteAllCommand::execute()
      */
     public function testExecute(): void
@@ -50,5 +50,17 @@ class DeleteAllCommandTest extends TestCase
         $this->assertOutputRegExp('/Driver: (Mysql|Postgres|Sqlite)/');
         $this->assertOutputContains('No backup has been deleted');
         $this->assertErrorEmpty();
+    }
+
+    /**
+     * @test
+     * @uses \DatabaseBackup\Command\DeleteAllCommand::execute()
+     */
+    #[WithoutErrorHandler]
+    public function testExecuteIsDeprecated(): void
+    {
+        $this->deprecated(function (): void {
+            $this->exec('database_backup.delete_all -v');
+        });
     }
 }