+++ tests/bootstrap.php
@@ -94,7 +94,7 @@ ConnectionManager::setConfig('test', ['url' => getenv('db_dsn')]);
 
 Configure::write('DatabaseBackup.connection', 'test');
 Configure::write('DatabaseBackup.target', TMP . 'backups' . DS);
-Configure::write('DatabaseBackup.mailSender', 'sender@example.com');
+
 /**
  * For Xampp
  */
@@ -116,10 +116,10 @@ echo 'Running tests for `' . (new BackupManager())->getDriverName() . '` driver 
 
 if (!function_exists('createBackup')) {
     /**
-     * Global function to create a backup file
+     * Global function to create a backup file.
+     *
      * @param string $filename Filename
      * @return string
-     * @throws \ReflectionException
      */
     function createBackup(string $filename = 'backup.sql'): string
     {
@@ -129,9 +129,9 @@ if (!function_exists('createBackup')) {
 
 if (!function_exists('createSomeBackups')) {
     /**
-     * Global function to create some backup files
+     * Global function to create some backup files.
+     *
      * @return string[]
-     * @throws \ReflectionException
      */
     function createSomeBackups(): array
     {